Moving Platforms
Moving Platform blueprint class. You are able to set the two end points of its track when placing it in the level. It should stop moving when hit with an actor with the tag "Projectile_Ice", and it should start moving when hit with an actor with the tag "Projectile_Fire". It has a customizable movement speed.